Run initialization step in on_apply if not joined yet
If the provider implementation does not guarantee that a view
is always applied before the first write set after joining to group,
the server initialization must be done in applying codepath.

The underlying reason for this is that wsrep-API does not have
joined_cb which would be called after the node has joined to
the cluster.

This change requires grabbing server mutex for each applied
write check to check the state.